requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Jinja2==3.1.2
Werkzeug==3.0.1

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# ML & Transformers
torch>=2.0.0
transformers>=4.30.0
peft>=0.6.0
accelerate>=0.20.0
safetensors>=0.4.0

# Time Series Forecasting
pytorch-forecasting>=1.0.0
pytorch-lightning>=2.0.0

# Data Processing
numpy==1.24.4
pandas==1.5.3
scikit-learn==1.3.2
pyarrow>=10.0.0

# Financial Data APIs
yfinance>=0.2.0
finnhub-python>=2.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# API Integration
requests>=2.31.0
newsapi-python>=0.2.7
